Fix TransitionConfig types link
The link explaining `TransitionConfig` is broken. I know it's not there in v5, so it's not going to matter soon, but a lot of people will probably still be stuck on v4 for a while, so it should still be useful. Hopefully making a PR to the `stable` branch is the right way to get this change up ¯\_(ツ)_/¯